requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
Django==4.2.4
djangorestframework==3.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
Werkzeug==2.3.7
django-cors-headers-4.2.0
